+++ src/constants/constants.go
@@ -1,0 +1,7 @@
+package constants
+
+const (
+	LinuxPlatformName   = "linux"
+	WindowsPlatformName = "windows"
+	DarwinPlatformName  = "darwin"
+)